MB-16181: Map SystemEvent to mcbp::systemevent
This commit maps the ep-engine SystemEvent enum value to mcbp ones.
Not all SystemEvent entries are things we replicate using dcp and
changes to SystemEvent shouldn't result in changes to the values
we transit.

Reviewed-on: http://review.couchbase.org/78640
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/dcp/producer.cc
@@ -601,7 +601,7 @@ ENGINE_ERROR_CODE DcpProducer::step(struct dcp_message_producers* producers) {
                     getCookie(),
                     s->getOpaque(),
                     s->getVBucket(),
-                    uint32_t(s->getSystemEvent()),
+                    s->getSystemEvent(),
                     *s->getBySeqno(),
                     {reinterpret_cast<const uint8_t*>(s->getKey().data()),
                      s->getKey().size()},
